python/requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
nemo_toolkit[asr]>=1.22.0
torch>=2.0.0
scipy>=1.10.0
